split ReadingInsightAccordeon into 2 components, fixed styling
